backend_ml/requirements.txt
fastapi~=0.115.4
uvicorn~=0.32.0
python-dotenv~=1.0.1
pymongo[srv]~=4.10.1
numpy~=2.0.2
pandas~=2.2.3
seaborn~=0.13.2
matplotlib~=3.9.2
scikit-learn~=1.5.2
fastcore~=1.7.20
colorama~=0.4.6
requests~=2.32.3
dnspython~=2.7.0
idna~=3.10
httpcore~=1.0.7
httpx~=0.27.2
anyio~=4.6.2.post1
h11~=0.14.0
jupyterlab-pygments~=0.3.0
pip~=23.2.1
attrs~=24.2.0
wheel~=0.41.2
pillow~=11.0.0
tornado~=6.4.2
jinja2~=3.1.4
exceptiongroup~=1.2.2
sniffio~=1.3.1
setuptools~=68.2.0
click~=8.1.7
ipython~=8.18.1
certifi~=2024.8.30
starlette~=0.41.2
pydantic~=2.9.2
urllib3~=2.2.3
annotated-types~=0.7.0
charset-normalizer~=3.4.0
PyYAML~=6.0.2
packaging~=24.2
zipp~=3.21.0
platformdirs~=4.3.6